Copy resources in classpath-defined order.
The current behavior copies the resources in such a way that, in the
case of conflicting filenames, the one with the latest timestamp wins.
This behavior results from the use of IO.copy with the default value for
the "overwrite" parameter (false), which will only overwrite a file if
the timestamp is later.

Instead, conflicts should be resolved according to the provided
classpath in the same way that the JVM would resolve such conflicts,
i.e., the class found via the earliest classpath entry wins.

See "Specification order" in:
http://docs.oracle.com/javase/1.5.0/docs/tooldocs/windows/classpath.html

To accomplish this, we copy resources over in *reverse* classpath order,
telling IO.copy to always overwrite.